chore: clean up lib/analysis
