--- src/lib/analysis/index.ts
@@ -1,2 +1,0 @@
-export * from './positionAnalysis'
-export * from './mistakeDetection'
